Thu, Nov 30, 2023, 10:53 PM -08:00

+++ calculator.c
@@ -6,6 +6,6 @@ int main(void)
     long x = get_int("x: ");
     long y = get_int("y: ");
 
-    float z = (float) x / (float) y;
+    double z = (double) x / (double) y;
     printf("%.20f\n", z);
 }
